backend/requirements.txt
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
expiringdict==1.2.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.4.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mysql-connector-python==8.0.32
ordered-set==4.1.0
packaging==23.1
protobuf==3.20.3
Pygments==2.14.0
PyJWT==2.6.0
python-dotenv==1.0.0
rich==13.3.3
six==1.16.0
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.15.0
